Fixed problem with quoting indented code blocks.
When a code block uses a tab character as prefix then it fails to work correctly when quoted. This is because expand_tabs merges spaces with tabs and ends up with '>   Code' when expanding '>\tCode'. This fails to be recognized as a code block. The current fix is to always replace a tab with 4 spaces. It'll work fine for email, but there may be problems in other scenarios.